api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "api for filmemo party game"
authors = ["platput <plato@techtuft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.94.1"
uvicorn = "^0.21.1"
python-dotenv = "^1.0.0"
google-cloud-firestore = "^2.10.0"
websockets = "^10.4"
numpy = "^1.24.2"
openai = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
